# c: vectorize table lookups