@indutny 2019-07-03
Example of generated code: https://gist.github.com/indutny/8047e46247ad58ca0c39bd2a8ac2549a