@indutny 2019-07-03
Had to publish `6.1.1` to resolve this. :facepalm: