@indutny 2019-07-03
I think it should build on Windows and all supported Unixes, but would really appreciate @bnoordhuis sign-off if they have time :wink: